capmc(bugfix): Fix case where reservations can be orphaned
When a Chassis or ComputeModule are powered off all components below
them have reservations taken out on them. That new reservation list was
not being sent back to the caller so it could properly release the
reservations when the power off operation was done. Resolution was to
return the list of components that got reserved so it could be passed to
the reservation release function.