capmc(bugfix) Fix filter query for On/Off when using Redfish
Description:
When using Redfish, CAPMC was basing HSM queries off of the Off or On
state as listed in HSM. This was not correct because of the chances that
the power state between hardware and HSM may not be in-sync. Fixed the
query to only honor the Off and On filter when using HSM to do the
query.